debug

+++ silnik_szachowy/funkcje.h
@@ -3,7 +3,7 @@
 #include <iomanip>
 #include <random>   
 #include <unistd.h>
-#include <unordered_map>
+#include <map>
 
 #define pb push_back
 #define st first
@@ -43,8 +43,8 @@ string pole(int a,int b,int c,int d);
 vector <string> mozliwe_ruchy(pozycja *poz);
 void zmiana_glebokosci(pozycja *poz);
 
-void los1(int a,int b,long long tab[12][64]);
-void los2(int a,int b,long long tab[13]);
+void los1(long long a,long long b,long long tab[12][64]);
+void los2(long long a,long long b,long long tab[13]);
 vector <int> pole_w_liczby(char bwp1,char bwp2);
 int Zobrist_hash_start(pozycja *poz);
 int Zobrist_hash_ruch(string ruch,pozycja *poz,int hash);
